This is a major refactoring which changes the traversal mechanism to use an explicit stack of attributes, rather than trying to pass rewriting results on the C stack.  The purpose of the refactoring is to allow traversals in the style of attribute grammars, in which a traversal can compute a number of attributes for each node.  In addition, simple traversals, such as visitors, no longer need to implement a complex reducer interface.
The type checking pass has been rewritten to use the new traversal mechanism,
and has been moved to TypedEvaluator.cpp.  Lazy substitutions for dependent
types now work properly, because the substitution can be stored within the
attribute for each term.

The lexical scope object is now a container for inherited attributes.

The SSA pass is temporarily disabled, until it can be ported to the new system.